refactor: centralize global styles
