--- app/loja/globals.css
@@ -1,47 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-/* ==========================
-   🎨 Variáveis de cor customizadas
-   ========================== */
-:root {
-  --background: #1e2019; /* eerie_black */
-  --foreground: #f9c5c7; /* cornell_red-900 */
-  --text-primary: #dcdcdc; /* Platinum */
-  --text-secondary: #ababab; /* jet-800 */
-  --accent: #b11116; /* cornell_red */
-}
-
-@media (prefers-color-scheme: dark) {
-  :root {
-    --background: #0c0d0a; /* eerie_black-200 */
-    --foreground: #f9c5c7; /* cornell_red-900 */
-    --text-primary: #dcdcdc; /* Platinum */
-    --text-secondary: #8c8c8c; /* jet-700 */
-    --accent: #e81920; /* cornell_red-600 */
-  }
-}
-
-body {
-  background: linear-gradient(
-    270deg,
-    #1e2019,
-    #2a1a1c,
-    #3b2a2b,
-    #1e2019,
-    #2a1a1c,
-    #1e2019
-  );
-  background-size: 500% 500%;
-  background-repeat: no-repeat;
-  animation: gradient-x 30s ease-in-out infinite;
-  color: var(--text-primary);
-  font-family: "Inter", Arial, Helvetica, sans-serif;
-  transition: background-color 0.s ease, color 0.9s ease;
-}
-
-/* 🧩 Input com estilo moderno e leve */
-.input-style {
-  @apply w-full px-4 py-3 rounded-lg border border-[#b3b3b3] bg-[#f0f0f0] text-sm text-eerie_black focus:outline-none focus:ring-2 focus:ring-cornell_red-600 transition;
-}
